Overhaul dataframe/series index alignment.
